risk-scoring-api/riskapi/scoring/serializers.py
from rest_framework import serializers
from .models import RiskAssessment

class RiskAssessmentSerializer(serializers.ModelSerializer):
    class Meta:
        model = RiskAssessment
        fields = ['purpose', 'data_sensitivity', 'region', 'processor_name']