risk-scoring-api/riskapi/requirements.txt
Django==4.2
djangorestframework==3.14.0
gunicorn==20.1.0
redis==4.5.5
celery==5.2.7